+++ packages/bip39/src/bip39.ts
@@ -1,8 +1,7 @@
-// This is a partial reimplementation of BIP39 in Deno: https://github.com/bitcoinjs/bip39
+// This is a partial reimplementation of BIP39 : https://github.com/bitcoinjs/bip39
 // We only use the default Wordlist (english)
-// import { Sha256 } from "https://deno.land/std@0.153.0/hash/sha256.ts";
 import { toHex } from "@lucid-evolution/core-utils";
-import { createHash } from "node:crypto";
+import { createHash, webcrypto } from "crypto";
 
 const Sha256 = () => createHash("sha256");
 
@@ -76,10 +75,12 @@ function randomBytes(size: number): Uint8Array {
       for (let generated = 0; generated < size; generated += MAX_BYTES) {
         // buffer.slice automatically checks if the end is past the end of
         // the buffer so we don't have to here
-        crypto.getRandomValues(bytes.slice(generated, generated + MAX_BYTES));
+        webcrypto.getRandomValues(
+          bytes.slice(generated, generated + MAX_BYTES),
+        );
       }
     } else {
-      crypto.getRandomValues(bytes);
+      webcrypto.getRandomValues(bytes);
     }
   }