+++ packages/lucid/src/tx-builder/internal/Stake.ts
@@ -1,16 +1,12 @@
 import { Lovelace, Redeemer, RewardAddress } from "@lucid-evolution/core-types";
 import * as TxBuilder from "../TxBuilder.js";
 import { Effect, pipe } from "effect";
-import {
-  ERROR_MESSAGE,
-  TxBuilderError,
-  TxBuilderErrorCause,
-} from "../../Errors.js";
+import { ERROR_MESSAGE, TxBuilderError } from "../../Errors.js";
 import * as CML from "@anastasia-labs/cardano-multiplatform-lib-nodejs";
 import { toPartial, toV1, toV2, validateAddressDetails } from "./TxUtils.js";
 
-export const stakeError = (cause: TxBuilderErrorCause, message?: string) =>
-  new TxBuilderError({ cause, module: "Stake", message });
+export const stakeError = (cause: unknown) =>
+  new TxBuilderError({ cause: `{ Stake: ${cause} }` });
 
 export const registerStake = (
   config: TxBuilder.TxBuilderConfig,
@@ -21,14 +17,16 @@ export const registerStake = (
       validateAddressDetails(rewardAddress, config.lucidConfig),
       Effect.andThen((address) =>
         address.type !== "Reward"
-          ? stakeError("InvalidCredential", "Address type must be Reward type.")
+          ? stakeError(ERROR_MESSAGE.MISSING_REWARD_TYPE)
           : Effect.succeed(address),
       ),
     );
 
     const stakeCredential = yield* pipe(
       Effect.fromNullable(addressDetails.stakeCredential),
-      Effect.orElseFail(() => stakeError("MissingStakeCredential")),
+      Effect.orElseFail(() =>
+        stakeError(ERROR_MESSAGE.MISSING_STAKE_CREDENTIAL),
+      ),
     );
 
     const credential =
@@ -55,14 +53,16 @@ export const deRegisterStake = (
       validateAddressDetails(rewardAddress, config.lucidConfig),
       Effect.andThen((address) =>
         address.type !== "Reward"
-          ? stakeError("InvalidCredential", "Address type must be Reward type.")
+          ? stakeError(ERROR_MESSAGE.MISSING_REWARD_TYPE)
           : Effect.succeed(address),
       ),
     );
 
     const stakeCredential = yield* pipe(
       Effect.fromNullable(addressDetails.stakeCredential),
-      Effect.orElseFail(() => stakeError("MissingStakeCredential")),
+      Effect.orElseFail(() =>
+        stakeError(ERROR_MESSAGE.MISSING_STAKE_CREDENTIAL),
+      ),
     );
 
     switch (stakeCredential.type) {
@@ -87,17 +87,12 @@ export const deRegisterStake = (
         const script = yield* pipe(
           Effect.fromNullable(config.scripts.get(stakeCredential.hash)),
           Effect.orElseFail(() =>
-            stakeError(
-              "MissingScript",
-              `No script found, script hash: ${stakeCredential.hash}, consider using attach modules`,
-            ),
+            stakeError(ERROR_MESSAGE.MISSING_SCRIPT(stakeCredential.hash)),
           ),
         );
         const red = yield* pipe(
           Effect.fromNullable(redeemer),
-          Effect.orElseFail(() =>
-            stakeError("MissingRedeemer", ERROR_MESSAGE.MISSING_REDEEMER),
-          ),
+          Effect.orElseFail(() => stakeError(ERROR_MESSAGE.MISSING_REDEEMER)),
         );
         switch (script.type) {
           case "PlutusV1": {
@@ -120,7 +115,7 @@ export const deRegisterStake = (
             break;
           }
           case "Native": {
-            yield* stakeError("NotFound");
+            yield* stakeError(ERROR_MESSAGE.INVALID_SCRIPT);
             break;
           }
         }
@@ -140,10 +135,7 @@ export const withdraw =
         validateAddressDetails(rewardAddress, config.lucidConfig),
         Effect.andThen((address) =>
           address.type !== "Reward"
-            ? stakeError(
-                "InvalidCredential",
-                "Address type must be Reward type.",
-              )
+            ? stakeError(ERROR_MESSAGE.MISSING_REWARD_TYPE)
             : Effect.succeed(address),
         ),
       );
@@ -154,7 +146,7 @@ export const withdraw =
             CML.Address.from_bech32(rewardAddress),
           ),
         ),
-        Effect.orElseFail(() => stakeError("MissingStakeCredential")),
+        Effect.orElseFail(() => stakeError(ERROR_MESSAGE.MISSING_STAKE_CREDENTIAL),),
         Effect.andThen((address) =>
           CML.SingleWithdrawalBuilder.new(address, amount),
         ),
@@ -162,7 +154,7 @@ export const withdraw =
 
       const stakeCredential = yield* pipe(
         Effect.fromNullable(addressDetails.stakeCredential),
-        Effect.orElseFail(() => stakeError("MissingStakeCredential")),
+        Effect.orElseFail(() => stakeError(ERROR_MESSAGE.MISSING_STAKE_CREDENTIAL),),
       );
 
       switch (stakeCredential.type) {
@@ -171,47 +163,41 @@ export const withdraw =
           break;
         }
 
-        case "Script": {
-          const script = yield* pipe(
-            Effect.fromNullable(config.scripts.get(stakeCredential.hash)),
-            Effect.orElseFail(() =>
-              stakeError(
-                "MissingScript",
-                `No script found, script hash: ${stakeCredential.hash}, consider using attach modules`,
+      case "Script": {
+        const script = yield* pipe(
+          Effect.fromNullable(config.scripts.get(stakeCredential.hash)),
+          Effect.orElseFail(() =>
+            stakeError(ERROR_MESSAGE.MISSING_SCRIPT(stakeCredential.hash)),
+          ),
+        );
+        const red = yield* pipe(
+          Effect.fromNullable(redeemer),
+          Effect.orElseFail(() => stakeError(ERROR_MESSAGE.MISSING_REDEEMER)),
+        );
+        switch (script.type) {
+          case "PlutusV1": {
+            config.txBuilder.add_withdrawal(
+              withdrawBuilder.plutus_script(
+                toPartial(toV1(script.script), red),
+                CML.Ed25519KeyHashList.new(),
               ),
-            ),
-          );
-          const red = yield* pipe(
-            Effect.fromNullable(redeemer),
-            Effect.orElseFail(() =>
-              stakeError("MissingRedeemer", ERROR_MESSAGE.MISSING_REDEEMER),
-            ),
-          );
-          switch (script.type) {
-            case "PlutusV1": {
-              config.txBuilder.add_withdrawal(
-                withdrawBuilder.plutus_script(
-                  toPartial(toV1(script.script), red),
-                  CML.Ed25519KeyHashList.new(),
-                ),
-              );
-              break;
-            }
+            );
+            break;
+          }
 
-            case "PlutusV2": {
-              config.txBuilder.add_withdrawal(
-                withdrawBuilder.plutus_script(
-                  toPartial(toV2(script.script), red),
-                  CML.Ed25519KeyHashList.new(),
-                ),
-              );
-              break;
-            }
-            case "Native": {
-              yield* stakeError("NotFound");
-              break;
-            }
+          case "PlutusV2": {
+            config.txBuilder.add_withdrawal(
+              withdrawBuilder.plutus_script(
+                toPartial(toV2(script.script), red),
+                CML.Ed25519KeyHashList.new(),
+              ),
+            );
+            break;
+          }
+          case "Native": {
+            yield* stakeError(ERROR_MESSAGE.INVALID_SCRIPT);
+            break;
           }
         }
       }
-    });
+    }});